# aragonite-vbox/api.yaml
swagger: 2.0
info:
  title: Aragonite VirtualBox Runner Plugin
  description: |
    The VirtualBox plugin creates VirtualBox machines to run tests in, and communicates with the internal operating
    system to provide run information and transmit information to reporters.
    
    By default, the plugin starts a new server on port `5720`, although the port can be changed in the configuration.
    
    Communication occurs over Socket communication.  Examples are given using [socket.io-client][] running inside a Node
    server.
    
    [socket.io-client]: https://www.npmjs.com/package/socket.io-client
  license:
    name: MIT
    url: https://github.com/rweda/aragonite/blob/master/LICENSE
schemes:
  - ws
host: aragonite:5720
...
paths:
...
  /{mac}/archive/:
    get:
      description: |
        Fetch the archive file associated with this machine.
      parameters:
        - name: mac
          type: string
          in: path
          required: true
          description: Provide the MAC address of the current machine.
      responses:
        '404':
          description: The MAC address wasn't found.
        '204':
          description: The MAC address was located, but no archive was associated with the run.
        '200':
          description: Provides the contents of the archive file provided to Aragonite.
...